requirements.txt
antiorm==1.2.1
anyio==3.6.2
api-display-purposes==0.0.3
art==5.8
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
clarifai==2.6.2
click==8.1.3
colorama==0.4.6
configparser==3.8.1
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cycler==0.11.0
db==0.1.1
decorator==4.4.2
Django==4.1.4
djangorestframework==3.14.0
drf-yasg==1.21.5
EasyProcess==1.1
emoji==2.2.0
et-xmlfile==1.1.0
fastapi==0.94.1
Flask==2.2.2
fonttools==4.38.0
future==0.18.2
geojson==2.5.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huepy==1.2.1
idna==2.10
imageio==2.24.0
imageio-ffmpeg==0.4.8
inflection==0.5.1
iniconfig==2.0.0
instabot==0.117.0
instaclient==2.9.20
instagram-private-api==1.6.0.0
instagrapi==1.16.17
instaloader==4.9.1
instapy==0.6.15
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==2.6.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.3
MeaningCloud-python==2.0.0
mock==5.0.1
MouseInfo==0.1.3
moviepy==1.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.5.1
path==16.6.0
Pillow==9.4.0
pluggy==1.0.0
plyer==2.1.0
proglog==0.1.10
protobuf==3.20.3
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.14.1
pydantic==1.9.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL==23.0.0
pyowm==3.3.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-resize-image==1.1.20
python-telegram-bot==20.0
pytweening==1.0.4
pytz==2022.6
pywhatkit==5.4
PyYAML==6.0
pyzbar==0.1.8
qrcode==7.3.1
regex==2022.10.31
requests==2.25.1
requests-toolbelt==0.10.1
resize-image==0.4.0
responses==0.22.0
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
schedule==1.1.0
seaborn==0.12.2
selenium==4.6.0
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.26.1
swagger-spec-validator==3.0.3
telebot==0.0.4
Telethon==1.25.4
termcolor==2.2.0
toml==0.10.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
types-toml==0.10.8.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.21.0
watchfiles==0.18.1
webdriver-manager==3.8.5
webdriverdownloader==1.1.0.3
websockets==10.4
Werkzeug==2.2.2
wikipedia==1.4.0
wsproto==1.2.0
xlrd==2.0.1
